% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-httr.R
\name{catch_errors}
\alias{catch_errors}
\title{Catches httr errors and prints them nicely}
\usage{
catch_errors(httr_response)
}
\description{
Catches httr errors and prints them nicely
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
