% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-httr.R
\name{catch_errors_connect_v1}
\alias{catch_errors_connect_v1}
\title{Catches Connect v1 httr errors and prints them nicely}
\usage{
catch_errors_connect_v1(httr_response)
}
\description{
Catches Connect v1 httr errors and prints them nicely
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
