% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{personal_access_token_available}
\alias{personal_access_token_available}
\title{Check personal_access_token availability}
\usage{
personal_access_token_available(verbose = TRUE)
}
\value{
logical
}
\description{
Check if a personal_access_token is available in \code{\link{squareupr}}'s internal
\code{.state} environment.
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
