% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{sq_auth}
\alias{sq_auth}
\title{Log in to Square}
\usage{
sq_auth(personal_access_token = NULL, token = NULL,
  app_id = getOption("squareupr.app_id"),
  app_secret = getOption("squareupr.app_secret"),
  callback_url = getOption("squareupr.callback_url"),
  cache = getOption("squareupr.httr_oauth_cache"), verbose = FALSE)
}
\arguments{
\item{personal_access_token}{Square}

\item{token}{optional; an actual OAuth 2.0 token object or the path to a valid token
stored as an \code{.rds} file}

\item{app_id, app_secret, callback_url}{the "App Id","App Secret", and "Callback URL" 
when using a connected app. NOTE: You must setup your own connected app and supply 
these arguments.}

\item{cache}{logical or character; TRUE means to cache using the default cache 
file \code{.httr-oauth-squareupr}, FALSE means don't cache. A string means use 
the specified path as the cache file.}

\item{verbose}{logical; do you want informative messages?}
}
\description{
Log in using a Personal Access Token or OAuth 2.0 authenticaion. OAuth does
not require sharing a token, but will require creating a connected app in the dashboard. 
Once you add the app id and secret, anytime you use \code{sq_auth} without a personal 
access token, then you will be directed to your default web browser, asked to sign 
in to your Square account, and to grant the connected app permission to pass data 
back to you on your behalf. By default, the oauth token credentials are cached 
in a file named \code{.httr-oauth-squareupr} in the current working directory.
}
\examples{
\dontrun{
# log in using personal access token
sq_auth(personal_access_token = "ThisIsATestToken") 

# log in using OAuth 2.0
# Via brower or refresh of .httr-oauth-squareupr
sq_auth()

# Save token and log in using it
saveRDS(.state$token, "token.rds")
sq_auth(token = "token.rds")
}
}
