% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sq_catch_null}
\alias{sq_catch_null}
\title{Convenience Function to Map NULL to NA}
\usage{
sq_catch_null(x)
}
\arguments{
\item{x}{object; to be checked if NULL that returns NA if it is NULL}
}
\value{
object
}
\description{
This function checks if a value is null and if so, returns NA. This is helpful 
when pulling information from lists and formatting to a \code{data.frame} structure 
where new rows cannot be NULL.
}
\examples{
\dontrun{
sq_catch_null(3)
sq_catch_null(NULL)
sq_catch_null(list(x=1, y=2))
}
}
