% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customers.R
\name{sq_create_customer}
\alias{sq_create_customer}
\title{Create Customer}
\usage{
sq_create_customer(input_data, verbose = FALSE)
}
\arguments{
\item{input_data}{\code{list}; data used for creating or updating an object.}

\item{verbose}{logical; do you want informative messages?}
}
\value{
\code{tbl_df} of the created customer record
}
\description{
Creates a new customer for a business, which can have associated cards on file.
}
\details{
You must provide at least one of the following values in your request 
to this endpoint:
\itemize{
  \item \code{given_name}
  \item \code{family_name}
  \item \code{company_name}
  \item \code{email_address}
  \item \code{phone_number}
 }
This endpoint does not accept an idempotency key. If you accidentally create a 
duplicate customer, you can delete it with the DeleteCustomer endpoint. Required 
permissions: \code{CUSTOMERS_WRITE}.
}
\examples{
\dontrun{
new_cust_data <- list(given_name = "API Test",
                      family_name = "API Test",
                      note = "This customer was created by the API as a test")
this_customer <- sq_create_customer(new_cust_data)
}
}
