% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items.R
\name{sq_create_item}
\alias{sq_create_item}
\title{Create Item}
\usage{
sq_create_item(location, input_data, verbose = FALSE)
}
\arguments{
\item{location}{character; the Square ID or name associated to a location. 
This must be an exact match to the ID or name as found using \link{sq_list_locations}.}

\item{input_data}{\code{list}; data used for creating or updating an object.}

\item{verbose}{logical; do you want informative messages?}
}
\value{
\code{tbl_df} of the created item record
}
\description{
Creates an item and at least one variation for it.
}
\details{
You must provide at the following fields in your request to this endpoint:
\itemize{
  \item \code{name}
  \item \code{variations}
 }
Required permissions: \code{ITEMS_WRITE}.
}
\examples{
\dontrun{
new_item_data <- list(name = "API Test Item",
                      variations = list())
this_item <- sq_create_item(location, 
                            new_item_data)
}
}
