% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customers.R
\name{sq_delete_customer}
\alias{sq_delete_customer}
\title{Delete Customer}
\usage{
sq_delete_customer(customer_id, verbose = FALSE)
}
\arguments{
\item{customer_id}{character; the Square ID assigned to each unique customer. 
This can be found using \link{sq_list_customers}}

\item{verbose}{logical; do you want informative messages?}
}
\value{
\code{tbl_df} of a single customer
}
\description{
Deletes a customer from a business, along with any linked cards on file.
}
\details{
Required permissions: \code{CUSTOMERS_WRITE}
}
\examples{
\dontrun{
sq_delete_customer(customer_id="ThisIsATestCustomerId")
}
}
