% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items.R
\name{sq_delete_item}
\alias{sq_delete_item}
\title{Delete Item}
\usage{
sq_delete_item(location, item_id, verbose = FALSE)
}
\arguments{
\item{location}{character; the Square ID or name associated to a location. 
This must be an exact match to the ID or name as found using \link{sq_list_locations}.}

\item{item_id}{character; the Square ID assigned to each unique item. 
This can be found using \link{sq_list_items}}

\item{verbose}{logical; do you want informative messages?}
}
\value{
\code{tbl_df} of a single item
}
\description{
Deletes an existing item and all item variations associated with it.
}
\details{
Required permissions: \code{ITEMS_WRITE}
}
\examples{
\dontrun{
sq_delete_item(item_id="ThisIsATestItemId")
}
}
