% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sq_extract_cust_groups}
\alias{sq_extract_cust_groups}
\title{Extract Group Membership from Customer Data}
\usage{
sq_extract_cust_groups(customer_data)
}
\arguments{
\item{customer_data}{\code{tbl_df} or \code{data.frame} containing an "id" and 
"groups" field}
}
\value{
a \code{tbl_df} of customers and their groups
}
\description{
This function reformats the groups field of a list of customers longways (i.e. 
one row per customer per group).
}
\examples{
\dontrun{
our_customers <- sq_list_customers()
cust_groups <- sq_extract_cust_groups(our_customers)
}
}
