% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customers.R
\name{sq_get_customer}
\alias{sq_get_customer}
\title{Get Customer}
\usage{
sq_get_customer(customer_id, verbose = FALSE)
}
\arguments{
\item{customer_id}{character; the Square ID assigned to each unique customer. 
This can be found using \link{sq_list_customers}}

\item{verbose}{logical; do you want informative messages?}
}
\value{
\code{tbl_df} of a single customer
}
\description{
Returns details for a single customer.
}
\details{
Required permissions: \code{CUSTOMERS_READ}
}
\examples{
\dontrun{
this_customer <- sq_get_customer(customer_id="ThisIsATestCustomerId")
}
}
