% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items.R
\name{sq_get_item}
\alias{sq_get_item}
\title{Get Item}
\usage{
sq_get_item(location, item_id, verbose = FALSE)
}
\arguments{
\item{location}{character; the Square ID or name associated to a location. 
This must be an exact match to the ID or name as found using \link{sq_list_locations}.}

\item{item_id}{character; the Square ID assigned to each unique item. 
This can be found using \link{sq_list_items}}

\item{verbose}{logical; do you want informative messages?}
}
\value{
\code{tbl_df} of a single item
}
\description{
Provides the details for a single item, including associated modifier lists and fees.
}
\details{
Required permissions: \code{ITEMS_READ}
}
\examples{
\dontrun{
this_item <- sq_get_item(location = location, 
                         item_id="ThisIsATestItemId")
}
}
