% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{sq_get_location}
\alias{sq_get_location}
\title{Get Location}
\usage{
sq_get_location(location)
}
\arguments{
\item{location}{character; the Square ID or name associated to a location. 
This must be an exact match to the ID or name as found using \link{sq_list_locations}.}

\item{verbose}{logical; do you want informative messages?}
}
\value{
\code{tbl_df} of a single location
}
\description{
Returns details for a single location.
}
\details{
Required permissions: \code{MERCHANT_PROFILE_READ}
}
\examples{
\dontrun{
# retrieve details using the location id
this_location1 <- sq_get_location(location = "ThisIsATestLocationId")

# retrieve details using the location name
this_location2 <- sq_get_location(location = "My Store Name")
}
}
