% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payments.R
\name{sq_get_payment}
\alias{sq_get_payment}
\title{Get Payment}
\usage{
sq_get_payment(location, payment_id, verbose = FALSE)
}
\arguments{
\item{location}{character; the Square ID or name associated to a location. 
This must be an exact match to the ID or name as found using \link{sq_list_locations}.}

\item{payment_id}{character; the Square ID assigned to each unique payment. 
This can be found using \link{sq_list_payments}}

\item{verbose}{logical; do you want informative messages?}
}
\value{
\code{tbl_df} of a single payment
}
\description{
Provides comprehensive information for a single payment.
}
\details{
Required permissions: \code{PAYMENTS_READ}
}
\examples{
\dontrun{
our_locations <- sq_list_locations()
our_payments <- sq_list_payments(location = our_locations$id[1], 
                                 begin=Sys.Date()-1, end=Sys.Date())
# return just one payment                                          
this_payment <- sq_get_payment(location = our_locations$id[1], 
                               payment_id = our_payments$id[1])
}
}
