% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categories.R
\name{sq_list_categories}
\alias{sq_list_categories}
\title{List Categories}
\usage{
sq_list_categories(location, verbose = FALSE)
}
\arguments{
\item{location}{character; the Square ID or name associated to a location. 
This must be an exact match to the ID or name as found using \link{sq_list_locations}.}

\item{verbose}{logical; do you want informative messages?}
}
\value{
\code{tbl_df} of categories
}
\description{
Lists all of a location's item categories.
}
\details{
Required permissions: \code{ITEMS_READ}
}
\examples{
\dontrun{
my_categories <- sq_list_categories(location)
}
}
