% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customers.R
\name{sq_list_customers}
\alias{sq_list_customers}
\title{List Customers}
\usage{
sq_list_customers(verbose = FALSE)
}
\arguments{
\item{verbose}{logical; do you want informative messages?}
}
\value{
\code{tbl_df} of customers
}
\description{
Lists a business's customers.
}
\details{
Required permissions: \code{CUSTOMERS_READ}
}
\examples{
\dontrun{
my_customers <- sq_list_customers()
}
}
