% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discounts.R
\name{sq_list_discounts}
\alias{sq_list_discounts}
\title{List Discounts}
\usage{
sq_list_discounts(location, verbose = FALSE)
}
\arguments{
\item{location}{character; the Square ID or name associated to a location. 
This must be an exact match to the ID or name as found using \link{sq_list_locations}.}

\item{verbose}{logical; do you want informative messages?}
}
\value{
\code{tbl_df} of discounts
}
\description{
Lists all of a location's discounts.
}
\details{
Required permissions: \code{ITEMS_READ}
}
\examples{
\dontrun{
my_discounts <- sq_list_discounts(location)
}
}
