% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fees.R
\name{sq_list_fees}
\alias{sq_list_fees}
\title{List Fees}
\usage{
sq_list_fees(location, verbose = FALSE)
}
\arguments{
\item{location}{character; the Square ID or name associated to a location. 
This must be an exact match to the ID or name as found using \link{sq_list_locations}.}

\item{verbose}{logical; do you want informative messages?}
}
\value{
\code{tbl_df} of fees
}
\description{
Lists all of a location's fees.
}
\details{
Required permissions: \code{ITEMS_READ}
}
\examples{
\dontrun{
my_fees <- sq_list_fees(location)
}
}
