% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items.R
\name{sq_list_items}
\alias{sq_list_items}
\title{List Items}
\usage{
sq_list_items(location, cursor = NULL, verbose = FALSE)
}
\arguments{
\item{location}{character; the Square ID or name associated to a location. 
This must be an exact match to the ID or name as found using \link{sq_list_locations}.}

\item{cursor}{character; A pagination cursor returned by a previous call to 
this endpoint. Typically this is not needed because functions in the package 
will automatically paginate results.}

\item{verbose}{logical; do you want informative messages?}
}
\value{
\code{tbl_df} of items
}
\description{
Provides summary information for all of a location's items.
}
\details{
Required permissions: \code{ITEMS_READ}
}
\examples{
\dontrun{
my_items <- sq_list_items(location)
}
}
