% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{sq_list_locations}
\alias{sq_list_locations}
\title{List Locations}
\usage{
sq_list_locations(verbose = FALSE)
}
\arguments{
\item{verbose}{logical; do you want informative messages?}
}
\value{
\code{tbl_df} of locations
}
\description{
Provides the details for all of a business's locations.
}
\details{
Most other Connect API endpoints have a required location_id path parameter. 
The id field of the Location objects returned by this endpoint correspond to 
that location_id parameter. Required permissions: \code{MERCHANT_PROFILE_READ}
}
\examples{
\dontrun{
my_locations <- sq_list_locations()
}
}
