% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{sq_list_modifiers}
\alias{sq_list_modifiers}
\title{List Modifiers.}
\usage{
sq_list_modifiers(location, verbose = FALSE)
}
\arguments{
\item{location}{character; the Square ID or name associated to a location. 
This must be an exact match to the ID or name as found using \link{sq_list_locations}.}

\item{verbose}{logical; do you want informative messages?}
}
\value{
\code{tbl_df} of modifiers.
}
\description{
Lists all of a location's modifier.
}
\details{
Required permissions: \code{ITEMS_READ}
}
\examples{
\dontrun{
my_modifiers <- sq_list_modifiers(location)
}
}
