% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transactions.R
\name{sq_list_transactions}
\alias{sq_list_transactions}
\title{List Transactions}
\usage{
sq_list_transactions(location, begin_time = Sys.Date() - 1,
  end_time = Sys.Date(), sort_order = c("DESC", "ASC"),
  cursor = NULL, verbose = FALSE)
}
\arguments{
\item{location}{character; the Square ID or name associated to a location. 
This must be an exact match to the ID or name as found using \link{sq_list_locations}.}

\item{begin_time}{Date or DateTime class; The beginning of the requested reporting 
period. The default value is one day prior at midnight local time (i.e. start of yesterday). 
If the value is a Date (no time component) the time is started at midnight of the date 
of the local timezone.}

\item{end_time}{Date or DateTime class; The end of the requested reporting period. 
The default value is today at midnight local time (i.e. start of today). If the 
value is a Date (no time component) the time is started at midnight of the date 
of the local timezone.}

\item{sort_order}{character; The order in which results are listed in the response 
(\code{ASC} for oldest first, \code{DESC} for newest first). The default value is \code{DESC}.}

\item{cursor}{character; A pagination cursor returned by a previous call to 
this endpoint. Typically this is not needed because functions in the package 
will automatically paginate results.}

\item{verbose}{logical; do you want informative messages?}
}
\value{
\code{tbl_df} of transactions
}
\description{
Lists transactions for a particular location.
}
\details{
Transactions include payment information from sales and exchanges and
refund information from returns and exchanges. Required permissions: \code{PAYMENTS_READ}
}
\examples{
\dontrun{
our_locations <- sq_list_locations()
yesterdays_transactions <- sq_list_transactions(our_locations$id[1])

sorted_transactions <- sq_list_transactions(our_locations$id[1], 
                                            begin_time = Sys.Date() - 1, 
                                            end_time = Sys.Date(), 
                                            sort_order = "ASC")
# specify the time range as datetimes: 
#   - Beginning April 6th, 2018 at 5PM EDT                                    
#   - Ending April 8th, 2018 at 8AM EDT                                  
begin <- as.POSIXct("2018-04-06 17:00:00", tz="America/New_York")
end <- as.POSIXct("2018-04-08 8:00:00", tz="America/New_York")
custom_time_range <- sq_list_transactions(our_locations$id[1], 
                                          begin_time = begin, 
                                          end_time = end)                                        
}
}
