% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sq_null_to_na}
\alias{sq_null_to_na}
\title{Convenience Function to Map \code{NULL} to \code{NA}}
\usage{
sq_null_to_na(x)
}
\arguments{
\item{x}{object; to be checked if NULL that returns NA if it is NULL}
}
\value{
object
}
\description{
This function checks if a value is null and if so, returns NA. This is helpful 
when pulling information from lists and formatting to a \code{data.frame} structure 
where new rows cannot be NULL.
}
\examples{
\dontrun{
sq_null_to_na(3)
sq_null_to_na(NULL)
sq_null_to_na(list(x=1, y=2))
}
}
