% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{sq_personal_access_token}
\alias{sq_personal_access_token}
\title{Return personal_access_token resulting from Basic auth routine}
\usage{
sq_personal_access_token(verbose = TRUE)
}
\arguments{
\item{verbose}{logical; do you want informative messages?}
}
\value{
character; a string of the sessionId element of the current authorized 
API session; otherwise NULL
}
\description{
Return personal_access_token resulting from Basic auth routine
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
