% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customers.R
\name{sq_search_customers}
\alias{sq_search_customers}
\title{Search Customers}
\usage{
sq_search_customers(cursor = NULL, limit = NULL, query = NULL,
  verbose = FALSE)
}
\arguments{
\item{cursor}{character; A pagination cursor returned by a previous call to 
this endpoint. Typically this is not needed because functions in the package 
will automatically paginate results.}

\item{limit}{integer; A limit on the number of results to be returned in a single 
page. The limit is advisory - the implementation may return more or fewer results. 
If the supplied limit is negative, zero, or is higher than the maximum limit of 
1,000, it will be ignored.}

\item{query}{list; A list containing \code{filter} and \code{sort} elements. 
Calling SearchCustomers without an explicit query parameter will return all 
customers ordered alphabetically based on \code{given_name} and \code{family_name}.}

\item{verbose}{logical; do you want informative messages?}
}
\value{
\code{tbl_df} of customers
}
\description{
Searches the customer profiles associated with a Square account. Calling SearchCustomers 
without an explicit query parameter returns all customer profiles ordered alphabetically 
based on given_name and family_name.
}
\details{
Required permissions: \code{CUSTOMERS_READ}
}
\examples{
\dontrun{
my_customers <- sq_search_customers()
}
}
