% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{sq_token}
\alias{sq_token}
\title{Return either the personal_access_token or OAuth access_token}
\usage{
sq_token()
}
\value{
character; an access token used for authorizing requests
}
\description{
Return either the personal_access_token or OAuth access_token
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
