% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customers.R
\name{sq_update_customer}
\alias{sq_update_customer}
\title{Update Customer}
\usage{
sq_update_customer(customer_id, input_data, verbose = FALSE)
}
\arguments{
\item{customer_id}{character; the Square ID assigned to each unique customer. 
This can be found using \link{sq_list_customers}}

\item{input_data}{\code{list}; data used for creating or updating an object.}

\item{verbose}{logical; do you want informative messages?}
}
\value{
\code{tbl_df} of the updated customer record
}
\description{
Updates the details of an existing customer.
}
\details{
The ID of the customer may change if the customer has been merged 
into another customer. Required permissions: \code{CUSTOMERS_WRITE}
}
\examples{
 
\dontrun{
updated_cust_data <- list(family_name = "API Test Test Test",
                          note = "This customer was updated by the API as a test.")
updated_customer <- sq_update_customer(customer_id="ThisIsATestCustomerId",
                                       updated_cust_data)
# you can confirm the update by retrieving the customer
this_customer <- sq_get_customer(updated_customer$id[1])
}
}
