% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items.R
\name{sq_update_item}
\alias{sq_update_item}
\title{Update Item}
\usage{
sq_update_item(item_id, input_data, verbose = FALSE)
}
\arguments{
\item{item_id}{character; the Square ID assigned to each unique item. 
This can be found using \link{sq_list_items}}

\item{input_data}{\code{list}; data used for creating or updating an object.}

\item{verbose}{logical; do you want informative messages?}

\item{location}{character; the Square ID or name associated to a location. 
This must be an exact match to the ID or name as found using \link{sq_list_locations}.}
}
\value{
\code{tbl_df} of the updated item record
}
\description{
Modifies the core details of an existing item.
}
\details{
If you want to modify an item's variations, use the Update Variation 
endpoint instead. If you want to add or remove a modifier list from an item, 
use the Apply Modifier List and Remove Modifier List endpoints instead. If you 
want to add or remove a fee from an item, use the Apply Fee and Remove Fee 
endpoints instead. Required permissions: \code{ITEMS_WRITE}.
}
\examples{
 
\dontrun{
updated_item_data <- list(name = "API Test Test Test Item")
updated_item <- sq_update_item(location, 
                               item_id="ThisIsATestItemId",
                               updated_cust_data)
# you can confirm the update by retrieving the item
this_item <- sq_get_item(updated_item$id[1])
}
}
