% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squareupr.R
\docType{package}
\name{squareupr}
\alias{squareupr}
\alias{squareupr-package}
\title{\code{squareupr} package}
\description{
An R package connecting to Square REST APIs using tidy principles
}
\details{
A package that connects to Square REST APIs and emphasizes the use of tidy 
data principles found in the tidyverse.

Additional material can be found in the 
\href{https://github.com/StevenMMortimer/squareupr}{README} on GitHub
}
