% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-httr.R
\name{VERB_n}
\alias{VERB_n}
\title{Generic implementation of HTTP methods with retries and authentication}
\usage{
VERB_n(VERB, n = 5)
}
\arguments{
\item{VERB}{function; an HTTP verb (e.g. GET, POST, etc.)}

\item{n}{integer; the number of retries}
}
\description{
Generic implementation of HTTP methods with retries and authentication
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
